formatting

+++ FlashContractV2/Flashloan.sol
@@ -3,33 +3,31 @@ pragma experimental ABIEncoderV2;
 
 import "@studydefi/money-legos/dydx/contracts/DydxFlashloanBase.sol";
 import "@studydefi/money-legos/dydx/contracts/ICallee.sol";
-import { KyberNetworkProxy as IKyberNetworkProxy } from '@studydefi/money-legos/kyber/contracts/KyberNetworkProxy.sol';
+import {KyberNetworkProxy as IKyberNetworkProxy} from "@studydefi/money-legos/kyber/contracts/KyberNetworkProxy.sol";
 
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
-import './IUniswapV2Router02.sol';
-import './IWeth.sol';
-
-//hacktoberfest2022
+import "./IUniswapV2Router02.sol";
+import "./IWeth.sol";
 
 contract Flashloan is ICallee, DydxFlashloanBase {
-    enum Direction { KyberToUniswap, UniswapToKyber } 
+    enum Direction {
+        KyberToUniswap,
+        UniswapToKyber
+    }
     struct ArbInfo {
         Direction direction;
-        uint repayAmount;
+        uint256 repayAmount;
     }
 
-    event NewArbitrage (
-      Direction direction,
-      uint profit,
-      uint date
-    );
+    event NewArbitrage(Direction direction, uint256 profit, uint256 date);
 
     IKyberNetworkProxy kyber;
     IUniswapV2Router02 uniswap;
     IWeth weth;
     IERC20 dai;
     address beneficiary;
-    address constant KYBER_ETH_ADDRESS = 0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE;
+    address constant KYBER_ETH_ADDRESS =
+        0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE;
 
     constructor(
         address kyberAddress,
@@ -38,11 +36,11 @@ contract Flashloan is ICallee, DydxFlashloanBase {
         address daiAddress,
         address beneficiaryAddress
     ) public {
-      kyber = IKyberNetworkProxy(kyberAddress);
-      uniswap = IUniswapV2Router02(uniswapAddress);
-      weth = IWeth(wethAddress);
-      dai = IERC20(daiAddress);
-      beneficiary = beneficiaryAddress;
+        kyber = IKyberNetworkProxy(kyberAddress);
+        uniswap = IUniswapV2Router02(uniswapAddress);
+        weth = IWeth(wethAddress);
+        dai = IERC20(daiAddress);
+        beneficiary = beneficiaryAddress;
     }
 
     // This is the function that will be called postLoan
@@ -55,52 +53,55 @@ contract Flashloan is ICallee, DydxFlashloanBase {
         ArbInfo memory arbInfo = abi.decode(data, (ArbInfo));
         uint256 balanceDai = dai.balanceOf(address(this));
 
-        if(arbInfo.direction == Direction.KyberToUniswap) {
-          //Buy ETH on Kyber
-          dai.approve(address(kyber), balanceDai); 
-          (uint expectedRate, ) = kyber.getExpectedRate(
-            dai, 
-            IERC20(KYBER_ETH_ADDRESS), 
-            balanceDai
-          );
-          kyber.swapTokenToEther(dai, balanceDai, expectedRate);
+        if (arbInfo.direction == Direction.KyberToUniswap) {
+            //Buy ETH on Kyber
+            dai.approve(address(kyber), balanceDai);
+            (uint256 expectedRate, ) = kyber.getExpectedRate(
+                dai,
+                IERC20(KYBER_ETH_ADDRESS),
+                balanceDai
+            );
+            kyber.swapTokenToEther(dai, balanceDai, expectedRate);
 
-          //Sell ETH on Uniswap
-          address[] memory path = new address[](2);
-          path[0] = address(weth);
-          path[1] = address(dai);
-          uint[] memory minOuts = uniswap.getAmountsOut(address(this).balance, path); 
-          uniswap.swapExactETHForTokens.value(address(this).balance)(
-            minOuts[1], 
-            path, 
-            address(this), 
-            now
-          );
+            //Sell ETH on Uniswap
+            address[] memory path = new address[](2);
+            path[0] = address(weth);
+            path[1] = address(dai);
+            uint256[] memory minOuts = uniswap.getAmountsOut(
+                address(this).balance,
+                path
+            );
+            uniswap.swapExactETHForTokens.value(address(this).balance)(
+                minOuts[1],
+                path,
+                address(this),
+                now
+            );
         } else {
-          //Buy ETH on Uniswap
-          dai.approve(address(uniswap), balanceDai); 
-          address[] memory path = new address[](2);
-          path[0] = address(dai);
-          path[1] = address(weth);
-          uint[] memory minOuts = uniswap.getAmountsOut(balanceDai, path); 
-          uniswap.swapExactTokensForETH(
-            balanceDai, 
-            minOuts[1], 
-            path, 
-            address(this), 
-            now
-          );
+            //Buy ETH on Uniswap
+            dai.approve(address(uniswap), balanceDai);
+            address[] memory path = new address[](2);
+            path[0] = address(dai);
+            path[1] = address(weth);
+            uint256[] memory minOuts = uniswap.getAmountsOut(balanceDai, path);
+            uniswap.swapExactTokensForETH(
+                balanceDai,
+                minOuts[1],
+                path,
+                address(this),
+                now
+            );
 
-          //Sell ETH on Kyber
-          (uint expectedRate, ) = kyber.getExpectedRate(
-            IERC20(KYBER_ETH_ADDRESS), 
-            dai, 
-            address(this).balance
-          );
-          kyber.swapEtherToToken.value(address(this).balance)(
-            dai, 
-            expectedRate
-          );
+            //Sell ETH on Kyber
+            (uint256 expectedRate, ) = kyber.getExpectedRate(
+                IERC20(KYBER_ETH_ADDRESS),
+                dai,
+                address(this).balance
+            );
+            kyber.swapEtherToToken.value(address(this).balance)(
+                dai,
+                expectedRate
+            );
         }
 
         require(
@@ -108,18 +109,17 @@ contract Flashloan is ICallee, DydxFlashloanBase {
             "Not enough funds to repay dydx loan!"
         );
 
-        uint profit = dai.balanceOf(address(this)) - arbInfo.repayAmount; 
+        uint256 profit = dai.balanceOf(address(this)) - arbInfo.repayAmount;
         dai.transfer(beneficiary, profit);
         emit NewArbitrage(arbInfo.direction, profit, now);
     }
 
     function initiateFlashloan(
-      address _solo, 
-      address _token, 
-      uint256 _amount, 
-      Direction _direction)
-        external
-    {
+        address _solo,
+        address _token,
+        uint256 _amount,
+        Direction _direction
+    ) external {
         ISoloMargin solo = ISoloMargin(_solo);
 
         // Get marketId from token address
@@ -138,7 +138,9 @@ contract Flashloan is ICallee, DydxFlashloanBase {
         operations[0] = _getWithdrawAction(marketId, _amount);
         operations[1] = _getCallAction(
             // Encode MyCustomData for callFunction
-            abi.encode(ArbInfo({direction: _direction, repayAmount: repayAmount}))
+            abi.encode(
+                ArbInfo({direction: _direction, repayAmount: repayAmount})
+            )
         );
         operations[2] = _getDepositAction(marketId, repayAmount);
